lxd-agent/metrics: Move Linux specific metrics functions to their own platform specific files
