+++ lxd-agent/metrics_linux.go
@@ -1,0 +1,56 @@
+//go:build linux
+
+package main
+
+import (
+	"bufio"
+	"bytes"
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/lxc/lxd/lxd/metrics"
+	"github.com/lxc/lxd/lxd/storage/filesystem"
+	"github.com/lxc/lxd/shared"
+)
+
+func getFilesystemMetrics(d *Daemon) (map[string]metrics.FilesystemMetrics, error) {
+	mounts, err := os.ReadFile("/proc/mounts")
+	if err != nil {
+		return nil, fmt.Errorf("Failed to read /proc/mounts: %w", err)
+	}
+
+	out := map[string]metrics.FilesystemMetrics{}
+	scanner := bufio.NewScanner(bytes.NewReader(mounts))
+
+	for scanner.Scan() {
+		fields := strings.Fields(scanner.Text())
+
+		// Skip uninteresting mounts
+		if shared.StringInSlice(fields[2], defFSTypesExcluded) || defMountPointsExcluded.MatchString(fields[1]) {
+			continue
+		}
+
+		stats := metrics.FilesystemMetrics{}
+
+		stats.Mountpoint = fields[1]
+
+		statfs, err := filesystem.StatVFS(stats.Mountpoint)
+		if err != nil {
+			return nil, fmt.Errorf("Failed to stat %s: %w", stats.Mountpoint, err)
+		}
+
+		fsType, err := filesystem.FSTypeToName(int32(statfs.Type))
+		if err == nil {
+			stats.FSType = fsType
+		}
+
+		stats.AvailableBytes = statfs.Bavail * uint64(statfs.Bsize)
+		stats.FreeBytes = statfs.Bfree * uint64(statfs.Bsize)
+		stats.SizeBytes = statfs.Blocks * uint64(statfs.Bsize)
+
+		out[fields[0]] = stats
+	}
+
+	return out, nil
+}
